move fn testnet_genesis to the new mod
