+++ runtime/vara/Cargo.toml
@@ -68,6 +68,7 @@ sp-api.workspace = true
 sp-authority-discovery.workspace = true
 sp-block-builder.workspace = true
 sp-consensus-babe.workspace = true
+sp-consensus-grandpa.workspace = true
 sp-core.workspace = true
 sp-externalities = { workspace = true, optional = true }
 sp-genesis-builder.workspace = true
@@ -199,6 +200,7 @@ std = [
 	"sp-authority-discovery/std",
 	"sp-block-builder/std",
 	"sp-consensus-babe/std",
+	"sp-consensus-grandpa/std",
 	"sp-core/std",
 	"sp-externalities",
 	"sp-inherents/std",
